Added post proccessing effects to smoothen Graphics
Changed objects that do not move(The Map Objects) to "Static" and than went to Window > Render > Occlusion Culling > Bake
RESULT: The objects themselves have a darker look and the scene looks nice but it ultimately did not fix the drop in performance

From the Top ribbon in Unity we went Edit > Project Settings > Quality and changed the Anti-aliasing from 4xMSAA to 8xMSAA
RESULT: Could not see a meaningful difference in performance

From the Top ribbon in Unity we went Oculus > Tools > OVRProfiler so we could monitor what the Headset was viewing and if that could give us any leads
RESULT: Once in use we found that the engine was going from 700 - 6000 draw calls(depending on where you are looking) which may be an issue. Looking further into how we can lower this.

From top ribbon went Window> Packages > Post Processing and we installed the latest version (2.1.6). This enabled us to add a component called "Post Processing Layer" to the main camera.
RESULT: From research online people have said that FXAA anti-aliasing works best for VR applications. The Post Processing layer gives us access to a FXAA anti-aliasing that the Project Settings did not have.

We kept the Map objects Static and also left the Occlusion Culling to Bake. We reverted the Anti-aliasing back to 4xMSAA in the project settings. We also Left the Post Processing Component on the Camera itself with a view of its options ticked.