Working with Facial expressions
Added Expression recipes to the Prefab McClients so they can blink and move there eyes around.
Also added a UMAMoodController Script to begin creating different facial expressions. Right now they currently have 5 expressions as follows:

     0 - Neutral
     1 - Happy
     2 - Sad
     3 - Mad
     4 - Surprised

Edited the MCQuestion script so that when a question is answered correctly the facial expression will change from Neutral to Happy, and when an incorrect answer it will go Neutral to Sad.

Played with the lightmapping settings of the computer desks to try and change the Dark choppy material. This has slightly worked but it will need to be looked further into. If you want this disabled just turn off the Directional Light in the gameobjects

Co-Authored-By: mwg223 <mwg223@users.noreply.github.com>