Alien Character
Imported a Werewolf UMA character race. Have begun playing with it's settings as it is from an older version of the UMA system. Looked into possible Race creation using Meshmixer but never fully delved into the compatibility of the program and the UMA system. From what i've seen I know that Blender program is compatible with UMA though.

Co-Authored-By: mwg223 <mwg223@users.noreply.github.com>