CashBoxClient revamp
Continued working on CashboxClient, fixed the issue with the Campus card being locked in position, as well as began working on the animation for the client to leave the area after thanking you. This works but still needs improvement's!

Also, tweaked the "WhiteSmoke" to be shorter and less intense from monitor destruction

Disabled Alien manager since I was not working on it at the time