Attempted UMA Fix
Changed some source code from the UMA scripts as theyare becoming obsolete.

Added Sphere Colliders and "UMA_Model" Tags to the Client models to try and set up a trigger event with the Sci-fi doors to open.

So far the trigger event hasn't worked as the Button collider is not detecting any collision event.